fix d2c_messaging.c: 261 (now 274)

+++ src/utils/d2c_messaging/src/d2c_messaging.c
@@ -271,7 +271,7 @@ static void DefaultIoTHubSendReportedStateCompletedCallback(int http_status_code
         goto done;
     }
 
-    for (int i = 0; i < message_processing_context->retryStrategy->httpStatusRetryInfoSize; i++)
+    for (size_t i = 0; i < message_processing_context->retryStrategy->httpStatusRetryInfoSize; i++)
     {
         ADUC_D2C_HttpStatus_Retry_Info* info = &message_processing_context->retryStrategy->httpStatusRetryInfo[i];
